requirements.txt
jira==3.8.0
python-dateutil==2.9.0
